pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "peptomatch"
version = "0.1.0"
description = "Genome-driven peptone recommendation system"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "openpyxl>=3.1",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "requests>=2.31",
    "pyyaml>=6.0",
    "streamlit>=1.28",
    "plotly>=5.17",
    "tqdm>=4.66",
]

[project.scripts]
peptomatch = "peptomatch.cli:main"
